+++ apps/omsorgspengerutbetaling-soknad/api-mock/mock-server.cjs
@@ -7,7 +7,7 @@ const busboyCons = require('busboy');
 
 require('dotenv').config();
 
-const barnMock = require('./mock-data/barn.json');
+const barnMock = require('./mock-data/ingen-barn.json');
 const søkerMock = require('./mock-data/soker.json');
 
 const server = express();