+++ apps/omsorgspengerutbetaling-soknad/e2e/playwright/utils/utfyllingUtils.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 import { Page, expect } from '@playwright/test';
 import dayjs from 'dayjs';
 import locale from 'dayjs/locale/nb.js';
@@ -43,9 +42,10 @@ const startSøknad = async (page: Page) => {
     await page.getByRole('button').getByText('Start søknad').click();
 };
 
-const fyllUtOmBarnMinstEttYngre13år = async (page: Page) => {
+const fyllUtOmBarnfyllUtOmBarnToUnder13år = async (page: Page) => {
     await page.getByRole('heading', { name: 'Om barn' });
-    await page.getByText('Ja, jeg bekrefter at jeg har dekket 10 omsorgsdager i år.').click();
+    await page.getByRole('group', { name: 'Har du fått ekstra omsorgsdager for et barn' }).getByLabel('Ja').click();
+    await page.getByRole('group', { name: 'Har du dekket de 10 første omsorgsdagene i år?' }).getByLabel('Ja').click();
     await page.getByTestId('typedFormikForm-submitButton').click();
 };
 
@@ -184,7 +184,7 @@ const kontrollerKvittering = async (page: Page) => {
 export const utfyllingUtils = {
     startSøknad,
     fyllUtVirksomhetDialog,
-    fyllUtOmBarnMinstEttYngre13år,
+    fyllUtOmBarnfyllUtOmBarnToUnder13år,
     fyllUtFraværSteg,
     lastOppLegeerklæring,
     fyllerUtArbeidssituasjonSteg,