+++ apps/omsorgspengerutbetaling-soknad/src/app/søknad/steps/oppsummering/components/DineBarnOppsummering.tsx
@@ -3,33 +3,58 @@ import intlHelper from '@navikt/sif-common-core-ds/src/utils/intlUtils';
 import { BarnType } from '@navikt/sif-common-forms-ds/src/forms/annet-barn/types';
 import { JaNeiSvar, SummaryBlock, SummaryList, SummarySection } from '@navikt/sif-common-ui';
 import { ApiBarn, RegistrertBarnTypeApi } from '../../../../types/søknadApiData/SøknadApiData';
+import { RegistrertBarn } from '../../../../types/RegistrertBarn';
+import { mapRegistrertBarnToApiBarn } from '../../../../utils/søknadsdataToApiData/getDineBarnApiDataFromSøknadsdata';
 
 interface Props {
     barn: ApiBarn[];
+    registrerteBarn: RegistrertBarn[];
+    harSyktBarn?: boolean;
+    harAleneomsorg?: boolean;
     harDekketTiFørsteDagerSelv?: boolean;
 }
 
-const DineBarnOppsummering = ({ barn, harDekketTiFørsteDagerSelv }: Props) => {
+const DineBarnOppsummering = ({
+    barn,
+    registrerteBarn,
+    harSyktBarn,
+    harAleneomsorg,
+    harDekketTiFørsteDagerSelv,
+}: Props) => {
     const intl = useIntl();
+    const registrerteBarnSomIkkeSkalSendesInnMenVises = registrerteBarn.map(mapRegistrertBarnToApiBarn);
     return (
         <SummarySection header={intlHelper(intl, 'step.oppsummering.dineBarn')}>
             <SummaryList
-                items={barn}
-                itemRenderer={({ identitetsnummer, navn, utvidetRett, type }: ApiBarn) => {
+                items={[...registrerteBarnSomIkkeSkalSendesInnMenVises, ...barn]}
+                itemRenderer={({ identitetsnummer, navn, type }: ApiBarn) => {
                     const fnr = identitetsnummer ? identitetsnummer : '';
-                    const harUtvidetRett = utvidetRett
-                        ? intlHelper(intl, 'step.oppsummering.dineBarn.listItem.utvidetRett')
-                        : '';
                     const barnType =
                         type !== BarnType.annet && type !== RegistrertBarnTypeApi.fraOppslag
                             ? intlHelper(intl, `step.oppsummering.dineBarn.listItem.årsak.${type}`)
                             : '';
-                    const punktum = type === RegistrertBarnTypeApi.fraOppslag && utvidetRett ? '.' : '';
-                    return <>{`${navn}${punktum} ${fnr} ${barnType} ${harUtvidetRett}`}</>;
+                    const punktum = type === RegistrertBarnTypeApi.fraOppslag ? '.' : '';
+                    return <>{`${navn}${punktum} ${fnr} ${barnType}`}</>;
                 }}
             />
-            {harDekketTiFørsteDagerSelv && (
-                <SummaryBlock header={intlHelper(intl, 'step.oppsummering.dineBarn.bekrefterDektTiDagerSelv')}>
+            {harSyktBarn !== undefined && (
+                <SummaryBlock
+                    header={intlHelper(
+                        intl,
+                        barn.length === 1
+                            ? 'step.dineBarn.utvidetRettSykdom.spm.ettBarn'
+                            : 'step.dineBarn.utvidetRettSykdom.spm',
+                    )}>
+                    <JaNeiSvar harSvartJa={harSyktBarn} />
+                </SummaryBlock>
+            )}
+            {harAleneomsorg !== undefined && (
+                <SummaryBlock header={intlHelper(intl, 'step.dineBarn.utvidetRettAleneomsorg.spm')}>
+                    <JaNeiSvar harSvartJa={harAleneomsorg} />
+                </SummaryBlock>
+            )}
+            {harDekketTiFørsteDagerSelv !== undefined && (
+                <SummaryBlock header={intlHelper(intl, 'step.dineBarn.bekrefterDektTiDagerSelv.spm')}>
                     <JaNeiSvar harSvartJa={harDekketTiFørsteDagerSelv} />
                 </SummaryBlock>
             )}