+++ apps/sif-ettersending/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-ettersending
 
+## 3.29.18
+
+### Patch Changes
+
+- Pakkeoppdateringer minor/patch
+- Updated dependencies
+  - @navikt/sif-common-amplitude@2.24.20
+  - @navikt/sif-common-formik-ds@1.7.14
+  - @navikt/sif-common-soknad-ds@10.13.14
+  - @navikt/sif-common-core-ds@9.15.5
+  - @navikt/sif-common-sentry@0.25.18
+
 ## 3.29.17
 
 ### Patch Changes