+++ packages/appstatus-react-ds/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @navikt/appstatus-react-ds
 
+## 2.17.23
+
+### Patch Changes
+
+- Pakkeoppdateringer minor/patch
+
 ## 2.17.22
 
 ### Patch Changes